Adds distributed computation of cv grids.

+++ analysis.py
@@ -1,10 +1,13 @@
 """Learns quark/gluon discrimination using one of several scikit-learn models.
 """
+from __future__ import division
 import argparse
 import sys
 import os
 import datetime
 import pickle
+import math
+import webbrowser
 
 from sklearn.ensemble import GradientBoostingClassifier
 from sklearn.naive_bayes import GaussianNB
@@ -13,9 +16,10 @@ from sklearn.metrics import classification_report
 from sklearn.metrics import confusion_matrix
 from sklearn.metrics import roc_curve, auc
 from sklearn.model_selection import train_test_split
-from sklearn.model_selection import GridSearchCV
 from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import StandardScaler
+from dask_searchcv import GridSearchCV
+from dask.distributed import Client
 import pandas as pd
 import numpy as np
 import matplotlib.pyplot as plt
@@ -23,7 +27,20 @@ import seaborn as sns
 
 import constants
 
-def get_data(modified=False):
+def calculate_jets(dataframe):
+  with np.errstate(divide='ignore', invalid='ignore'):
+    dataframe['dispersion'] = dataframe.apply(lambda r: np.sqrt(np.sum(np.square(r['trackPt']))) / np.sum(r['trackPt']), axis=1)
+    dataframe.loc[~np.isfinite(dataframe['dispersion']), 'dispersion'] = 0.0
+  def width(row):
+    deltaPhi = np.absolute(row['jetPhi']-row['trackPhi'])
+    deltaPhi = np.minimum(deltaPhi, 2*math.pi - deltaPhi)
+    deltaEta = row['jetEta']-row['trackEta']
+    deltaR = np.sqrt(np.square(deltaPhi) + np.square(deltaEta))
+    return np.sum(np.multiply(row['trackPt'], deltaR)) / row['jetPt']
+  dataframe['width'] = dataframe.apply(width, axis=1)
+  return dataframe.drop(columns=['trackPt', 'trackEta', 'trackPhi', 'trackCharge', 'towerE', 'towerEem', 'towerEhad', 'towerEta', 'towerPhi'])
+
+def get_data(modified=False, recalculate=False, max_events=0):
   if modified:
     pickle_file = os.path.join(constants.DATA_PATH, 'modified_data.pickle')
     quarks_path = constants.MODIFIED_QUARKS_PATH
@@ -32,24 +49,29 @@ def get_data(modified=False):
     pickle_file = os.path.join(constants.DATA_PATH, 'standard_data.pickle')
     quarks_path = constants.STANDARD_QUARKS_PATH
     gluons_path = constants.STANDARD_GLUONS_PATH
-  # Try to load the data from a pickle file first, otherwise use root numpy.
-  try:
-    with open(pickle_file, 'rb') as fh:
-      df_quarks, df_gluons = pickle.load(fh)
-  except (IOError, OSError) as e:
+  if recalculate or not os.path.exists(pickle_file):
     # Normally, imports should be at the top of the module. However, root_numpy
     # has undesirable side-effects when it is imported, to such an extent that
     # it breaks argparse when it is imported. As such, it is not imported until
     # absolutely necessary, in an attempt to limit the damage it does.
     from root_numpy import root2array
-    df_quarks = pd.DataFrame(root2array(quarks_path, branches=constants.BRANCH_NAMES))
-    df_gluons = pd.DataFrame(root2array(gluons_path, branches=constants.BRANCH_NAMES))
+    df_quarks = calculate_jets(pd.DataFrame(root2array(quarks_path)))
+    df_gluons = calculate_jets(pd.DataFrame(root2array(gluons_path)))
     with open(pickle_file, 'wb+') as fh:
       pickle.dump((df_quarks, df_gluons), fh)
+  else:
+    with open(pickle_file, 'rb') as fh:
+      df_quarks, df_gluons = pickle.load(fh)
   df_quark_labels = pd.Series(np.zeros(df_quarks.shape[0], dtype=np.uint8))
   df_gluon_labels = pd.Series(np.ones(df_gluons.shape[0], dtype=np.uint8))
-  X = pd.concat([df_quarks, df_gluons])
-  y = pd.concat([df_quark_labels, df_gluon_labels])
+  if max_events > 0:
+    X = pd.concat([df_quarks[:max_events], df_gluons[:max_events]])
+    y = pd.concat([df_quark_labels[:max_events], df_gluon_labels[:max_events]])
+  else:
+    X = pd.concat([df_quarks, df_gluons])
+    y = pd.concat([df_quark_labels, df_gluon_labels])
+  X.reset_index(drop=True, inplace=True)
+  y.reset_index(drop=True, inplace=True)
   return X, y
 
 def plot_roc(model, X, y, plot_path, title, color):
@@ -100,26 +122,33 @@ def print_grid_performance(model):
 def main():
   parser = argparse.ArgumentParser(description='Analyze quark/gluon separation using a particular machine learning model.')
   parser.add_argument('-m', '--model', choices=['NB', 'NN', 'GBRT'], required=True, help='Type of machine learning model to use.')
+  parser.add_argument('--recalculate', action='store_true', help='If set, recalculate data from root files, instead of using pickled data.')
+  parser.add_argument('--max_events', default=0, type=int, help='The maximum number of standard quark/gluon events to use in training (max_events <= 0 indicates to use all of them)')
   args = parser.parse_args()
-  X_standard, y_standard = get_data(modified=False)
-  X_modified, y_modified = get_data(modified=True)
-  if args.model == 'GBRT':
-    param_grid = {'learning_rate': [0.1, 0.4, 0.7, 1.0], 'n_estimators': [50, 100, 150], 'max_depth': [2, 3, 4], 'subsample': [0.5, 0.75, 1.0]}
-    model = GridSearchCV(GradientBoostingClassifier(), param_grid, cv=3, verbose=50)
+  X_standard, y_standard = get_data(modified=False, recalculate=args.recalculate, max_events=args.max_events)
+  X_modified, y_modified = get_data(modified=True, recalculate=args.recalculate)
+  if args.model == 'GBRT' or args.model == 'NN':
     grid = True
+    client = Client('localhost:8786')
+    webbrowser.open('http://localhost:8787')
+  else:
+    grid = False
+  if args.model == 'GBRT':
+    param_grid = {'learning_rate': [0.01, 0.1], 'n_estimators': [10, 100, 1000], 'max_depth': [2, 3, 4], 'subsample': [0.1, 0.5, 1.0]}
+    model = GridSearchCV(GradientBoostingClassifier(), param_grid, scheduler=client)
   elif args.model == 'NB':
     model = GaussianNB()
-    grid = False
   elif args.model == 'NN':
     pipeline = Pipeline([('scale', StandardScaler()), ('nn', MLPClassifier())])
     param_grid = {'nn__alpha': (10.0 ** -np.arange(1, 7)).tolist()}
-    model = GridSearchCV(pipeline, param_grid, cv=3, verbose=50)
-    grid = True
+    model = GridSearchCV(pipeline, param_grid, scheduler=client)
   X_eval, y_eval = train(model, X_standard, y_standard, print_report=True)
   if grid:
     print_grid_performance(model)
   timestamp = '{:%Y%m%d%H%M%S}'.format(datetime.datetime.now())
   os.mkdir(timestamp)
+  with open(os.path.join(timestamp, 'args.txt'), 'w+') as fh:
+    fh.write(str(args))
   standard_plot_path = os.path.join(timestamp, 'standard_qg_roc.png')
   modified_plot_path = os.path.join(timestamp, 'modified_qg_roc.png')
   plot_roc(model, X_eval, y_eval, standard_plot_path, 'Receiver Operating Characteristic (Standard Data)', 'orange')
